+++ changelog.md
@@ -22,5 +22,6 @@
 
 ## Next
 
+* Run jshint on tests
 * Disable Travis dependency cache
 * Update README badges to use shields.io