+++ changelog.md
@@ -22,6 +22,7 @@
 
 ## Next
 
+* Test code coverage & Travis sends to coveralls
 * Run jshint on tests
 * Disable Travis dependency cache
 * Update README badges to use shields.io