+++ changelog.md
@@ -22,6 +22,7 @@
 
 ## Next
 
+* Update dev dependencies in line with Sequelize v3.2.0
 * Travis runs tests with Sequelize v3 and v2
 * Test code coverage & Travis sends to coveralls
 * Run jshint on tests