+++ changelog.md
@@ -22,6 +22,7 @@
 
 ## Next
 
+* Support for Sequelize v3.x.x
 * Update dev dependencies in line with Sequelize v3.2.0
 * Travis runs tests with Sequelize v3 and v2
 * Test code coverage & Travis sends to coveralls